(ajaxhook) Major cleanup of ajaxhook logic: use dataFilter callback
This should solve IE-related problems of #9.
In IE11, it was impossible to make built-in fields writable
via Object.defineProperty, so we couldn't overwrite xhr.responseText.

Because MediaWiki extensions use "mediawiki.Api" module, we can wrap
around its ajax() method and rewrite the response on this level.
This should be browser-independent.