(visualeditor) Load preload.ve.js early, so that VisualEditor won't call unmodified requestParsoidData() on slow hardware
A quick fix for a possible race condition when VisualEditor calls requestParsoidData()
before our script (preload.ve.js) has a chance to wrap around this function.

This fix is primarily for testing in Sauce Labs (see issue #13).
When not running in a tiny virtual machine, this race condition wasn't observed.