(ajax) Preliminary version of Ajax handling on Special:Moderation (disabled by default) See issue #15.
This feature is not ready yet (e.g. "diff" or "block" links are not yet implemented).
To test it (for development purposes only), set $wgModerationUseAjax = true;

Note that:
- format of Special:Moderation is not modified. It still works without JavaScript.
- Ajax handlers are installed by [ext.moderation.special.ajax.js].
- result of (asynchronous) Ajax action is indicated by icon (for now 16x16) near every row,
because adding text message like "Approved" to rows could change their position,
and we don't want the user to accidentally click on the wrong link because of that.
- possible icons:
	untouched (white square) - user hasn't done any action to this line.
	processing - clicked on some action link, not completed yet.
	approved/rejected - successfully completed action.
	error - has a tooltip which shows the error message.

TODO:
- actions: show, block/unblock, merge
- extensive UX testing
- Selenium tests