(ajax) Disable Ajax for show/merge links (for now) See issue #15.
1. merge action ("Resolve edit conflict") requires edit form to be opened.
Even though we could display a modal dialog for it on Special:Moderation,
it's probably not going to be as convenient (or mobile-friendly) as opening it
in a different tab.

2. When moderators check Special:Moderation, they usually open many diff links
in tabs and then examine them.
Displaying show action ("diff") via Ajax (on Special:Moderation itself)
would ultimately be less convenient for the moderator, as page space is limited
and only one edit will be loaded at once.

Above-mentioned considerations may be countered in the future,
e.g. if we develop a multi-tab viewer widget for "diff" links.

For now, these links are simply opened as usual (not via Ajax).