(tests,TDD) Test doEditContent() between explicit begin() and rollback()
This test is to find a solution for commit() issue in ModerationEditHooks.
See issue #26.

We are trying to reliably preserve the newly added row in 'moderation' table,
when the caller of doEditContent() throws MWException, causing a DB rollback.

This test currently fails due to internal sanity checks in commit()
which are designed to specifically prevent what we are trying to do
(interfering with someone else's explicit transaction).

Possible solution is to allow the rollback and repeat the changes AFTER it.