(feature) Improve protection of successfully queued edits from rollback() See issue #26.
This reliably ensures that newly added row in 'moderation' table won't be lost
if the caller of doEditContent() throws MWException, causing a DB rollback.

Previous solution was to call commit() prematurely after insert()/update();
it was a bad idea for the following reasons:
(1) it prevented atomicity of the current transaction,
(2) it needed different handling with DBO_TRX or explicit transation on/off,
(3) it had to workaround sanity checks inside commit().

New solution is to detect database rollback and re-insert the queued edit
after the rollback is completed.

Utility class RollbackResistantQuery is provided with high-level methods
	RollbackResistantQuery::insert()
	RollbackResistantQuery::update()