(feature) Support MultiUpload-like extensions in MediaWiki 1.28+
Here we use UploadVerifyUpload hook (appeared in MediaWiki 1.28)
to become compatible with extensions that perform multiple uploads
in one request (alternatives of Special:Upload).

For MediaWiki 1.27, we make a polyfill of UploadVerifyUpload,
which uses the old UploadVerifyFile hook. This doesn't offer support
of above-mentioned upload extensions (because UploadVerifyFile
doesn't know things like image description, and we have to get them
from $wgRequest, which is not used by MultiUpload extensions).