(tests) Refactoring of phpunit testsuite: prepare to support internal invocation
When we need to test API/HTTP requests, we do a real request via the network.
However, someone might want to run tests where network is not available,
and they would prefer testing via internal invocation.

NOTE: testing the real requests is obviously better, because when using mocks,
you can sometimes miss errors due to imperfection of the emulation layer.
So we still want to support "real requests" mode (and recommend to use it).

This change provides an abstraction layer ModerationTestsuiteEngine,
which currently has only one subclass: ModerationTestsuiteRealHttpEngine.
An internal invocation version may then be subclassed as an optional mode.