(tests) InternalInvocation: isolate tested API calls via pcntl_fork()
Some parts of MediaWiki are written with the assumption that only one request
is executed per PHP invocation.

For example, ModerationApproveHooks install additional hooks during approval.
However, these hooks aren't removed when they are no longer needed, because
(1) determining when to remove them is messy,
(2) this is unnecessary: in production PHP exits after modaction=approve.

When testing via internal invocation, we run several tests in one PHP process,
so it's important to isolate such "global side-effects" of the tested code
from the tests that run later.

New convenience method forkAndRun() is provided.
It runs the test in the child process and returns the result to the parent.

NOTE: we don't want to use native PHPUnit feature @runInSeparateProcess,
because we need to provide non-serializable globals to the child process,
so pcntl_fork() is the logical choice.