(feature,DB-SCHEMA-CHANGE) Preserve tags set by Extension:AbuseFilter
When filters of AbuseFilter try to assign tags to the edit,
these tags will be remembered by Moderation. When the moderator later
approves the edit, the tags will be applied as usual.

Important: this change modifies the SQL table used by Moderation.
The feature will only be enabled after running update.php.