(refactor) Provide ModerationDatabaseEntry class
We move things like "save change by user A in page B into the database"
away from ModerationEditHooks, because this code may be used when intercepting
other events (like renaming pages).

The syntax should probably be improved, but the code is more readable already.

The idea is for the caller to:
1) create $entry - an object of type ModerationDatabaseEntry,
2) pass whatever information is known via methods like $entry->setSummary(),
all of them being optional.
3) let $entry->queue() calculate the necessary database fields.