(typo) Don't show Approve link on modaction=show for null edits
Note: we still show Approve link in Pending folder of Special:Moderation,
as checking "is this a null edit?" for each row would be impractically slow.

Adding field like mod_null_edit (0/1) is also impractical, because non-null
edit can become a null edit (or vise versa) because of the edits of users
who can skip moderation, therefore this field would need to be recalculated
after every edit. Not really worth it.