(1.31,VisualEditor) Clear autosave of VisualEditor after the edit
Starting with MediaWiki 1.31, VisualEditor keeps its own "pending revision"
in mw.storage.session, restoring it when the edit form is reopened.
This leads to our preloading method (from [preload.ve.js]) not being called.

When the edit is successfully queued for moderation, we must clear the autosave
storage, because it doesn't restore the edit summary (our preload does).