(compat) Return $dbr->timestamp(), will be needed for PostgreSQL
MediaWiki uses different format of timestamps if PostgreSQL is used.

Moderation doesn't support PostgreSQL yet (database patches are
for MySQL only, we need to write them for PostgreSQL too),
but it may support PostgreSQL in the future.

Note: we don't use $dbr->timestamp() in ApproveHook::install() call,
because $row->timestamp is already in the correct format
for whatever database is currently being used.