(feature) Experimental: intercept page moves (disabled by default)
If $wgModerationInterceptMoves=true, page moves will be intercepted.
WARNING: this feature not yet ready, because:

1) page moves should be displayed as non-edits on Special:Moderation.

2) modaction=show doesn't support them yet (should show "new title").

3) being automoderated alone shouldn't be enough to bypass moderation of moves,
because they can be used for hard-to-revert vandalism (there should be another
right, e.g. 'skip-move-moderation').

However, this option allows to write/run move-related tests on testwiki.

Note: once this feature is refined, $wgModerationInterceptMoves will be removed
(and the moves will be intercepted by default).