(tests) Benchmarks: use unique Titles, don't touch existing database
It's an unexpected behavior when benchmarks clean the database.

Instead we will prepend the Title of tested pages with a unique time-based
prefix which will be different for each benchmark run.
This is similar to how our Selenium tests handle inability to clean the DB.

Here we completely unbind ModerationBenchmark class from ModerationTestsuite.