(performance) Removed unneeded Revision::newFromId() - WikiPage already has it
When EntryEdit::doApprove() was resolving the edit conflict automatically,
call to Revision::newFromId( $latest ) always caused a SQL query.

However, $latest is the rev_id of the latest revision in WikiPage $page,
and $page most likely already has this revision loaded.
So if we call $page->getContent() instead, there won't be an extra SQL query.