Revert "(performance) Cache the results of isModerationBlocked() check"
This cache doesn't provide any noticeable benefit, because the check
is only called when saving the edit (which alone requires some queries
to DB_MASTER), and even if not cached,
it consists of one superfast well-indexed SQL query.

At the same time, this cache makes the code difficult to test,
and third-party command line scripts may have problems invalidating it
when $wgMainCacheType is CACHE_ACCEL or something.

This reverts commit 0c29cc3e36a3fc5352abcc9c2aa6c0aa9f6be75d.