(tests) TestsuiteSelfTest: skip this test by default (not needed)
This test is only needed when debugging ModerationTestsuiteEngine,
it doesn't check the work of Extension:Moderation itself.

Note: use "tests/phpunit/phpunit.php --group Utility" to run it anyway.

This avoids the need to setup a web server in Travis testing, etc.
(where it's easier to use CliEngine, not RealHttpEngine).