(tests) CliEngine: non-uopz solution to intercepting header() calls: MockAutoLoader
Because "uopz" PHP extension is rather unstable (often segfaults),
here is a pure-PHP solution to intercepting header() in PHP CLI.

We provide an autoloader method MockAutoLoader::autoload().
Every time it is called, it finds the PHP file that MediaWiki autoloader
would use and preprocesses it, applying the following replacement:

header() -> efModerationTestsuiteMockedHeader()

... and saves modified file under another name (".filename.mocked.php").

After that $wgAutoloadClasses is modified, so that MediaWiki autoloader
would load the modified file instead of the original.