(tests) .travis.yml: temporarily allow Selenium failures due to T199393
See https://phabricator.wikimedia.org/T199393 for details.

Short story is,
when our Selenium tests try to create multiple test accounts simultaneously,
MediaWiki gets itself into a database deadlock (unrelated to Moderation).

A quick solution would be to reduce maxInstances to 1 in wdio.conf.sauce.js,
but do we want to wait 25 minutes for the results we can get in 3 minutes?

Correct solutions would be:
(1) investigate T199393 and fix it in MediaWiki core,
(2) precreate test accounts BEFORE the testsuite and pass them to the tests.