(tests) Ensure that afterTestHook() is called while the browser is still open
This should fix a Selenium error "Stale element reference" which was noticed
after a completely successful testsuite run (159/159 correct assertions).

The call to afterTestHook() from util/miscFunctions.js is the only thing that
happens after all assertions. Previously it was called from after() hook in
wdio.conf.js, but it's possible that the browser has already been closed.

We should call it earlier, in the afterTest() hook of wdio.conf.js.