(tests) Provide browser.precreatePageAsync() to make Selenium tests faster
Some tests (in notify.js and visualeditor.js) need an existing page.
Previously we waited for it to be created in before() hook.
By that time the browser is already open and can be used for other checks.

Now before() will only initiate the "create page" operation, and then we
immediately proceed with other tests. Since only the last checks (in notify
and VisualEditor tests) will need this page, there is a performance bonus
due to not keeping the Selenium server idle (especially with SauceLabs).