(tests) Remove duplicate tests: already in QueueTest and InterceptTest
With the development of decoupled testsuite (QueueTest/InterceptTest),
the following tests became superfluous:

1) ModerationWatchTest - these checks are fully covered by QueueTest.
2) testInterceptEdit() - already checked in InterceptTest.
3) testModeratorNotAutomoderated() - the fact that edit is intercepted
is verified by InterceptTest. The remainder of the test is unnecessary,
because approval procedure is not affected by the user groups.

TODO: many other tests have checks that are duplicating QueueTest and
RenderTest, but we can't remove them,
because the decoupled testsuite doesn't check modactions yet.