(tests) Decoupled testsuite: preliminary ModerationActionTest
ModerationActionTest is the third part of the decoupled testsuite.
It does the following:

1) pending edit/upload/move is created (as in RenderTest),
2) URL of tested modaction is downloaded,
3) consequences are asserted (output of modaction, database state
afterwards, for approvals - whether the change was indeed made, etc.)

Because (1) is the same for RenderTest and ActionTest,
we move this logic into PendingChangeTestSet (subclass of TestSet),
which is then subclassed by testsets of RenderTest and ActionTest.

Moreover, assertion "do all fields of the database row match expected?"
is the same for QueueTest and ActionTest, so we move this logic
into TestSet::assertRowEquals() convenience method.