(tests) Check (1) timestamps being recent, (2) 'moderation_block' table
Here we make tests more thorough:

1) assertTimestampIsRecent() will check that the timestamp of test edit
is not too far from "now".
This ensures that *_timestamp fields won't have incorrect values.

- QueueTest checks mod_timestamp
- ActionTest checks mb_timestamp and log_timestamp

2) ActionTest: "is user modblocked?" will always be checked, even if
the action is not block-related. If "expectModblocked" is not set
to true/false, we assert that the block status shouldn't have changed.

3) ActionTest now checks the fields of 'moderation_block' row
(mb_user, mb_by, mb_by_text).