(UX,i18n) Report "success" when trying to (un)block an (un)blocked user
If moderator B tries to modblock user A, but A is already modblocked,
then there is no reason to report this situation as an error,
because the desired outcome has been reached anyway.

It can happen if B clicked "Mark as spammer" twice or if another
moderator just blocked A.

Instead we will say "User A was successfully marked as spammer"
(because A has indeed been marked as spammer, even if B didn't do it).

The following interface messages are removed (no longer needed):
- moderation-block-fail
- moderation-unblock-fail