(cosmetic) Replace wfMessage() with $context->msg() whenever possible
Use of wfMessage() makes the code less convenient to test,
because it always uses the global RequestContext.

Now we will respect the local $context if/when it is available.

Note: we don't replace wfMessage() in ModerationNewChange, because
this class is logically not dependent on the current user,
plus its messages are always used with inContentLanguage().