(feature) Preliminary version of modaction=editchange (disabled by default)
This is intended for wikis that use Moderation for pre-publish review
of new pages.
If $wgModerationEnableEditChange=true in LocalSettings.php,
then the moderator can modify text/summary of the pending change before
approving it. This way unreviewed revisions are never published.

NOTE: this is disabled by default, because moderator can overwrite the
original text of the change, which can't be undone (pending changes
don't have an edit history).

If $wgModerationEnableEditChange is disabled, modaction=editchange will
only show the wikitext of pending change, but won't allow to modify it.