(editchange) Remove readonly mode of action=editchange (not needed)
If $wgModerationEnableEditChange (default: false) is not set to true,
then modaction=editchange should be completely disabled,
with "no such action" error returned.
Showing the edit form in readonly mode is counter-intuitive.

While it's possible that someone would want to see wikitext of pending
change without actually editing it, this should be in a separate action
(because this feature is not specific to pre-publish review).