(tests) Fix flaky testApproveDeletedUser (loginAs after deleting user)
Test testApproveDeletedUser() sometimes fails in loginAs($moderator),
where action=clientlogin rejects a fresh login token as invalid.

The possible explanation is:
1) the test calls loginAs($moderator) immediately after deleting
$unprivilegedUser by direct modification of the database.
2) before (1) we were logged in as $unprivilegedUser.
3) it's possible that due to some cache or something, $unprivilegedUser
is still used when generating a loginToken in loginAs().
4) however, by the time this loginToken is used, it is already invalid,
even though loginAs() just obtained it.

A possible workaround is $t->logout() before deleting the user.
This way we can be sure that loginAs() was called as anonymous user
and the loginToken wouldn't be dependent on $unprivilegedUser.