(cosmetic) Ignore ?thumb=1 when viewing non-images (e.g. OGG files)
This change eliminates an ugly exception in modaction=showimg&thumb=1
when this URL corresponds to non-image (e.g. OGG audio file).
Now "thumb=1" will be ignored (full-size file will be downloaded).

Apparently $file->transform() on OGG file returns a non-error result,
but its getStoragePath() returns false (so we can't stream it).

Note: this situation never happened in production, because
modaction=show uses links (not thumbnails) for non-images.

We are only fixing this for consistency (to have the same behavior
as when thumbnail creation failed for any other reason).