(tests) .travis.yml: skip PHP 5.6 tests on 'phpunit' branch
Branch "phpunit" is used for developing the PHPUnit testsuite.
We don't really need PHP 5.6 compatibility checks here (they are only
used with MediaWiki 1.27, which is already tested with PHP 7.1),
because incompatibility of testsuite with PHP 5.6 (even if it appears
and gets merged into the master branch) won't affect production,
allowing us to fix it after the merge.