(tests,TDD) ActionTest: add "modaction=approve of the upload" test
This test currently fails, because modaction=approve on uploads doesn't
set "revid" parameter in LogEntry (revid is simply not yet known at the
moment when this LogEntry is created).

This should probably be fixed by SQL UPDATE on the 'logging' table
when revid becomes known (after DeferredUpdates of performUpload()).

Note: this issue doesn't affect:
1) normal edits
2) reuploads
3) moves ("approve-move" log entry shouldn't have "revid")