(bug) LogEntry "moderation/approve" should record "revid" for uploads
Previously "revid" parameter of these log entries was null,
because "revid" is the ID of new revision on description page of the
newly uploaded image, but this description page hasn't been created yet
(its creation is delayed by DeferredUpdates) when we inserted LogEntry.

We can't really delay inserting LogEntry (it should be in the same
transaction as the main changes, so that no error would result in
the change being approved without the LogEntry), but we can fix
this entry in the "logging" table later. We do this in FileUpload hook
(description page already exists when this hook is called).