(tests) ActionTest: test null edits, avoid two uploads in one second
This is a follow-up change for "test edit/upload/move on existing page"
in ActionTest.

We add "nullEdit" flag to PendingChangeTestSet (to emulate situation
when the old text hasn't changed). This is also used to upload a file
without any description.

We also make sure that reupload doesn't happen on the same second as
the upload (MediaWiki differentiates archive versions of images by
adding timestamp to their name, so it doesn't support two versions
of the same file with exactly the same timestamp).