(tests) PHPUnit testsuite: log only the relevant part of HTML responses
If we dump the whole HTML of the responses, the log becomes so large
that we can't print it after Travis runs (exceeds output length limit).

However, most tests only check the title tag and $html->getMainText(),
so we can log only these parts of HTML, omitting the navigation, etc.