(tests) Skip "ReadOnly mode" tests if CACHE_MEMCACHED is unavailable
When the wiki is in ReadOnly mode, CACHE_DB (which is commonly used
for $wgSessionCacheType due to its persistence) won't be available,
so the only option is CACHE_MEMCACHED.

ActionTest will now ensure that $wgSessionCacheType uses Memcached.

Furthermore, availability of Memcached server will be tested, and
if Memcached is not running, then ReadOnly tests will be skipped.

(with no CACHE_DB and CACHE_MEMCACHED, there is just no way to login,
and these tests require user to be logged in as moderator)