(bug) Default value of mod_page2_namespace must be 0, not empty string
This fixes newly noticed test failures (that didn't exist a week ago)
in MediaWiki 1.27 (old LTS).

The field mod_page2_namespace is INT, so attempting to insert empty
string into it would result in the following error:
	Incorrect integer value: ''

NOTE: this never caused any issues in the past,
because MediaWiki DB abstraction layer mitigates this somehow.

E.g. the following works in MediaWiki 1.31/1.32
(and until this week, it worked in MediaWiki REL1_27 as well):

1) CREATE TABLE test ( id INT NOT NULL DEFAULT 0, name VARCHAR(100) )
2) $dbw->insert( 'test', [ 'id' => '', 'name' => 'wow' ] )
3) ... and the following row is created with no error: id=0, name=wow.

Anyway, we shouldn't rely on MediaWiki fixing this for us,
we should just use explicit 0 for mod_page2_namespace.