(updater) Fix for ModerationVersion not being updated by web updater
This fixes issue #46 and issue #44.
It was noticed that markDbAsUpdated() doesn't work correctly when using
the web updater (instead of the usual update.php in the terminal).

The cause is that $wgExtensionCredits is not populated in web updater,
so we can't learn the current version of Moderation from it.

Impact: Moderation believed that the DB schema version is old,
so some newer features (such as intercepting page moves) were disabled.

Solution is to get the version from extension.json.