(tests) Remove duplicate test: already covered by LogFormatterTest
This test is removed for two reasons:

1) it's a duplicate: testApproveAll() already checks (via the API)
that the logentry has indeed been created,
and LogFormatterTest checks that such entries are formatted properly.

2) it became flaky in MediaWiki 1.33,
because Special:Log is ordered ONLY by timestamp (without log_id),
meaning that the latest log entry may sometimes not be on top,
if there is another log entry with the same timestamp.

This limitation of Special:Log can affect ApproveAll test,
since ApproveAll creates N approve entries and 1 approveall entry,
and they are likely to have the same timestamp.