(tests) PHPUnit testsuite: mark RealHttpEngine as deprecated
RealHttpEngine is no longer usable (since MW 1.33),
because MediaWiki is actively trying to isolate the testsuite from the
real database, and RealHttpEngine requires access to the real database.

The workaround in ModerationTestsuiteEngine::escapeDbSandbox(),
which helped RealHttpEngine since MW 1.28, is no longer working.

It's very sad that this "testsuite sandbox" thing is mandatory,
as RealHttpEngine was the cleanest approach to blackbox testing.