(tests) Travis: don't run SauceLabs tests on master branch (too flaky)
Selenium testing on SauceLabs became ridiculously unreliable.
Failures happen for tests that always succeed locally, and tests that
succeeded in SauceLabs in the past (with no changes in tested code).

The current theory is that the virtual machine of SauceLabs is so slow
that very unlikely race conditions (unrelated to Moderation itself) are
appearing in VisualEditor and MobileFrontend (which are very complex,
and making VE/MF "easy to test" is outside the scope of our testsuite).

At this point, seeing "a test has failed" from SauceLabs is no longer
useful, because with 99% certainty it's a false positive.
Double-checking each video is even slower than doing the test manually.

Until the test flakiness is somehow fixed,
there is no point to run SauceLabs tests from .travis.yml.