(tests) QueueTest: run upload tests with and without Storage migration
Running the tests on a completely empty database was not enough to test
the uploads, because migrateFromPerUploaderStashes() is always called
if [[User:ModerationUploadStorage]] doesn't exist,
and this method alters the test consequences we are supposed to check.

To ensure that migration didn't affect the result, upload-related tests
will be run with and without precreating that system user
(who owns the UploadStash).