(tests) Ensure that PHPUnit testsuite always uses "qqx" pseudo-language
When testing anonymous uploads in QueueTest, it was noticed that only
logged-in users were using "qqx" language. Anonymous users were using
the default language of the wiki (in this case, English).

Because of that, NonApiBot couldn't find "(moderation-image-queued)" in
the response, because English translation was shown instead.

Solution is to set "qqx" globally as a content language of the wiki
when running the testsuite.

This also means that engine subclasses now MUST implement setMwConfig(),
which was optional before (and is not supported by RealHttpEngine).
However, CliEngine (the only engine that works with modern MediaWiki)
supports it, and RealHttpEngine hasn't been working since MW 1.28,
so this can be safely ignored.