(bug) Default value of mod_stash_key must be NULL, not empty string
This fixes a minor regression bug.
(introduced when old code was replaced by ModerationNewChange class)

Previously mod_stash_key was NULL (and it should be NULL) for edits,
because "Approve all" uses "mod_stash_key IS NULL" as sorting
criteria to ensure that uploads are approved BEFORE edits.

This way if a non-automoderated user uploads B.png and includes B.png
into the article A, article A won't have a redlink after both this edit
and this upload are approved by "Approve all".

(if edit in article A was approved earlier, the file B.png wouldn't
exist yet when the Parser was processing the wikitext of A,
so the Parser would generate a redlink to [[File:B.png]])