(tests) Improve error reporting of ModerationTestSet::assertRowEquals()
Previously a mismatch on 1 field of the row resulted in an immediate
test failure, and that prevented us from knowing if other fields are
correct or not, as well as seeing the whole picture of what happened.

Now "expected values of fields" and "actual values of fields" will be
compared with one assertEquals() on these arrays (sorted by key).