(tests) Completely clear Memcached between subsequent PHPUnit tests
It was discovered that Memcached is not properly cleaned between tests,
and a certain test was failing due to reliance on the stale cached data
that was stored in Memcached during the previous test.

Solution is to exlicitly clean Memcached at the beginning of each test
by sending "flush_all" command directly to the Memcached server.