(tests) PHPUnit: don't clear Memcached keys related to other wikis
This is needed for potentially running multiple phpunit.php in threads,
where one thread would work with database DB1 and another with DB2.

Because $wgDBname is included into the memcached keys, we can list
the contents of memcached via LRU crawler
and then delete only what's needed.

As simple as it is, "flush_all" can't be used with parallel testings,
as some tests depend on Memcached to retain data during the test,
e.g. NotifyModeratorTest or readonly tests.