(postgres) Fix PostgreSQL incompatibility in ModerationVersionCheck
With PostgreSQL, the trick with adding a row to page_props table (with
nonexistent pp_page=-1) causes a foreign key constraint error.

The method getDbUpdatedVersionUncached() has been rewritten.
It now detects the changes in the schema that happened between versions
of Moderation, and learns the version from that.

Furthermore, checks are sorted "most recent changes first",
so the wiki with the most recent schema would only need one check.