(tests) Reset PostgreSQL sequences before tests, fix constraint errors
When clearing a table, it's important to reset the associated sequence,
because MediaWiki uses PostgreSQL sequences to emulate $dbw->insertId()
(MySQL function which doesn't exist in PostgreSQL),
and if the sequences are not reset, then the value of insertId() won't
be correct. (leading to failures in User::createNew(), etc.)

Additionally, several extra tables (e.g. 'log_search') are now cleared,
even though they are not used by the testsuite, because they have unique
and foreign key constraints, and leaving those tables filled with values
from previous test would cause SQL errors.

Additionally, commit() for those "truncate table" operations was moved
before the attempts to create new users,
because creating new users was failing in PostgreSQL due to unique key
constraints (related to rows that we just deleted, but not committed).

Additionally, a check "was User::createNew() successful?" was is added
to createTestUser() - without it error message was very non-intuitive.
(such error was triggered due to PostgreSQL above-mentioned constaint)