(postgres) PostgreSQL support: fix incorrect invocation of $dbw->upsert
This affects the following:
1) $dbw->upsert() in lib/ModerationNewChange.php
2) $dbw->replace() in tests/benchmarks/ModerationBenchmark.php

Parameter $uniqueFields was passed to $dbw->upsert() incorrectly:
if you supply an array of fields, upsert() joins them in OR conditions,
but we need AND condition.
To get AND condition, all fields should be passed as 1 element
within the $uniqueFields[] array, as expected by upsert().

Previously (when PostgreSQL wasn't supported) this error has no impact,
because MySQL supports upsert natively (INSERT..ON DUPLICATE),
and it was not using $uniqueFields parameter at all.