(tests) Add sanity checks to CliEngine: "are we logged in as expected?"
When investigating the parallel testing issues, an odd session-related
issue (that very rarely was noticed in the past, and was thought to be
not actual) has been detected with nearly reproducible frequencies.

The following happens:
1) testsuite logins via api.php?action=clientlogin as "User:User_1"
(note that "User 1" is a moderator)
2) API responds: clientlogin: { status: "PASS", username: "User 1" }
3) testsuite makes a usual GET request to download Special:Moderation,
and gets a permissions error (indicating that this request was not
executed on behalf of "User 1" - a session was lost or something).

Apparently CliEngine has some underlying issues with session handling.

To troubleshoot this,
1) an extra log was added "what CliEngine runs on behalf of whom",
2) CLI-invoked script will throw an exception if the User object
in the global context was noticed to be different from expected.