(tests) Make parallel PostgreSQL tests work with $prepopulateDbCache
The reason why parallel PostgreSQL tests where failing was because of
the incorrect order of INSERT statements in prepopulateDb(),
leading to incorrect values in mwuser.user_id field.

This didn't affect MySQL, because for MySQL we explicitly included
the correct value of user_id into the INSERT query.
This was skipped for PostgreSQL, because the sequence of mwuser table
would only be incremented if primary key field (user_id) is omitted.

This also eliminates arbitrary test failures that we were trying
to troubleshoot via TestsuiteEngine::getExpectedUser().
Note: getExpectedUser() will remain as a suitable sanity check
against similar regression bugs.