(phan) Fix PhanTypeMismatchArgumentNullable - false positive
Phan complains because it thinks that $newContent can be null for
non-wikitext contents (which don't support sections), but this doesn't
happen during normal operations, because EditPage throws an exception
when it finds "section=something" parameter for a non-wikitext page.

It's not even a concern if someone calls doEditContent() directly,
because we get $section from EditFilter hook (not from $request),
and this hook is called by EditPage.