(tests) PHPUnit testsuite: explore potential way to fix RealHttpEngine
RealHttpEngine hasn't been working since MediaWiki 1.28
(where "DB sandboxing" was added to MediaWiki testsuite runner script),
but there may be a way to fix it by using SandboxWikiSettings.php,
which would be included into LocalSettings.php and would
switch the wiki to unittest-cloned database for non-CLI invocations
of PHP (i.e. when using PHP-FPM or mod_php).

This way both phpunit.php and results of HTTP requests would be working
with the same underlying database.

Note: this is incomplete and needs further testing.

Note: RealHttpEngine is not necessary (CliEngine works really well),
but if it can be fixed, it might be nice to do so.
Its implementation is much shorter and simpler than that of CliEngine.