(tests) MockConsequenceManager::mockResult() should require class name
mockResult() must require the knowledge of "which Consequence to mock".

This would ensure that we won't return the mocked value for the wrong
Consequence if new Consequence gets added or if their order changes.

Note: $mockedResults in ActionsHaveConsequencesTest::getConsequences()
was simplified on purpose (previously it accepted multiple results,
now it only accepts 1 class name and 1 result). This is not a typo.