(lib) Provide WatchCheckbox class to keep track of "Watch this" fields
This improves performance of move/upload API calls by fixing the
following minor issue: affected pages were always Unwatched when
moved/uploaded NOT via UI of Special:Movepage and Special:Upload.

The cause was that "Watch this page" checkbox is exclusive to those
special pages, but Moderation does Watch/Unwatch in the hooks which
intercept the move/upload, and those hooks are called when moving and
uploading not via UI too.

Absence of the checkbox field in $request means its value is "false",
therefore Unwatch was always performed.

Similar problem didn't exist for normal edits (non-uploads, non-moves),
because ModerationEditHooks::$watchthis was set in EditFilter hook,
and this hook is only called when EditPage object (that has a checkbox)
exists and can be analyzed.

Now a similar approach is used for Move/Upload too.
The variable $watchthis is moved into a separate class WatchCheckbox,
and SpecialPageBeforeExecute hook will detect checkboxes if we are
on Special:Movepage or Special:Upload.
If we are not on such page, $watchthis remains "null", which means
that neither Watch nor Unwatch are needed. (false meant "must Unwatch")