(lib) Fix return value of InsertRowIntoModerationTable in PostgreSQL
It was noticed that $dbw->insertId() is incorrect (doesn't correspond
to mod_id of updated row) due to non-native $dbw->upsert() being used
for PostgreSQL by MediaWiki.
(it returns insertId() from noop INSERT IGNORE instead of insertId()
from UPDATE, which is what we need, but can't obtain)

Note: this return value isn't really used anywhere except tests
(not since ModerationNewChange::upload() method was implemented),
so it's a shame that it requires an extra SELECT query.

However, returning incorrect row number would be undocumented behavior,
and we are still fixing this for consistency.

Note: MySQL is unaffected by this change. (it uses native UPSERT)

Note: for PostgreSQL, performance impact is negligible,
as this is an indexed SELECT query for a newly inserted row,
and it only happens during a slow non-readonly action (editing a page).