(refactor) Provide ModerationApproveHook::destroySingleton()
The purpose of this change is to be able to "forget" ApproveHook tasks
when they are no longer needed.

This is primarily for tests, where leftover ApproveHook tasks from the
previous test shouldn't modify any environment (e.g. static properties)
that will be used by other tests later.

Implementation:
1) we move all "static" properties of ApproveHook into non-static ones.
2) ApproveHook itself is now a singleton. See ApproveHook::singleton().
3) ApproveHook can be reset to empty state (along with any properties
such as $tasks) by calling destroySingelton().
4) unit tests (but not isolated blackbox tests) call destroySingleton()
in tearDown() of any test that does approvals.