(lib) ApproveHook: don't expect timestamp to be already converted
Values of revision.rev_timestamp are converted via $db->timestamp(),
because timestamp format is different in PostgreSQL.

However, ApproveHook shouldn't expect this in "timestamp" field that is
passed to it, because it's not intuitive to the caller
that this parameter goes into the database field.

Instead, ApproveHook should encode it with $db->timestamp() itself.