(lib) Call ApproveHook::doUpdate after the last edit OR move
Preamble: this DeferredUpdate is queued multiple times (with $useCount
that is incremented every time), and is only executed once.

Previously we queued it only in PageContentSaveComplete hook
(to ensure that doUpdate() gets executed after ALL the edits are made),
but this doesn't cover the page moves: they don't cause this hook.

Therefore we'll additionally queue this update in TitleMoveComplete
hook (which is called after each move).