(bug) Fix non-updated rev_timestamp in redirects after approving a move
When writing detailed unit tests of InstallApproveHookConsequence,
the following issue was noticed:

When renaming page A to B, newly created redirect A might not exist
at the moment when RecentChange_save hook is invoked.
Because of that, its rev_id couldn't have been properly determined,
and rev_timestamp of this revision wasn't updated by ApproveHook.

Solution is to queue this update from within TitleMoveComplete hook,
because at this point the redirect A is already known to exist.