(lib) Fix ApproveHook ignoring timestamp on edit/move of the same page
This fixes handling of the following situation:
1) user A made edit in page B at timestamp=100,
2) user A renamed page B at timestamp=200,
3) moderator used "Approve all" on user A at timestamp=300.

Previously renaming (2) didn't apply rev_timestamp=200, because it was
earlier than the timestamp of previous revision, which is currently 300
(because this check happens before ApproveHook had a chance to replace
this 300 with 100).

Now we'll take into account that ApproveHook plans to replace this 300
with 100, and use it in "should timestamp from (2) be ignored?" check.

Note: these checks are ordered by timestamp (from oldest to newest),
so if in some case we need to ignore (1) setting the timestamp to 100,
then this ignored 100 won't be used when checking (2).